requirements.txt
nextcord 
discord.py
colorama
requests
